requirements.txt
python-telegram-bot==20.0
beautifulsoup4==4.12.0
requests==2.31.0
selenium==4.15.0
pandas==2.0.3
sqlalchemy==2.0.0
psycopg2-binary==2.9.0  # Nếu dùng PostgreSQL trên Render